List/pyproject.toml
[tool.poetry]
name = "list"
version = "0.1.0"
description = ""
authors = ["Lavderim Pacuku"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
django-axes = "^7.0.0"
django-extensions = "^3.2.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
